Reorganizing the repo into controllers and services
